Save requestor instance ID for internal follow up messages
After the pairing sequence, additional follow-up frames are generated
to exchange keys using the SKDA attribute. Store the requestor's
instance ID and sub_pub_id for reuse with the same peer. Also
update requestor instance ID and sub_pub_id only if non-zero.

During pairing verification for unsolicited publisher, PASN frame is
the first frame received by the publisher and requestor instance ID
is not known and the peerHandle is invalid. So don't update the IDs
if it is zero.

CRs-Fixed: 4227876
Change-Id: Ib3eb7c1378b81428318dc139c2cab593720acffe